+++ changelog.md
@@ -24,6 +24,7 @@
 * Made it possible for the admin to disable new signups.
 * Reduced the amount of queries when using the link row field.
 * Respect the date format when converting to a date field.
+* Added a field type filename contains filter.
 
 ## Released (2021-02-04)